requirements.txt
Django
djangorestframework
celery
redis
django-celery-beat